+++ wayland-server/src/native_lib/display.rs
@@ -70,7 +70,8 @@ impl DisplayInner {
                 handle
                     .insert_source(evtsrc, move |_, _| {
                         ffi_dispatch!(WAYLAND_SERVER_HANDLE, wl_event_loop_dispatch, evl_ptr, 0);
-                    }).unwrap(),
+                    }).map_err(Into::<::std::io::Error>::into)
+                    .unwrap(),
             );
 
             Rc::new(RefCell::new(DisplayInner {